This closes #66. Move util.h into HelperFunctions.
